authorization for countries done

+++ app/Http/Controllers/CountryController.php
@@ -12,14 +12,19 @@ use Illuminate\Http\Request;
 class CountryController extends Controller
 {
 
+    private $_class;
+
     private CountryService $_service;
     public function __construct(CountryService $service)
     {
         $this->_service = $service;
+        $this->_class = $this->_service->model()::class;
     }
 
     public function all(Request $request) : CountryCollection
     {
+        $this->authorize("viewCountries", $this->_class);
+
         $pages = $request->get("pages", 10);
         $language = $request->get("language", app()->getLocale());
         $model = $this->_service->model();
@@ -28,6 +33,8 @@ class CountryController extends Controller
 
     public function create(Request $request) : CountryResource
     {
+        $this->authorize("createCountry", $this->_class);
+
         $data = $request->all();
         $country = $this->_service->create($data);
         return new CountryResource($country);
@@ -36,6 +43,8 @@ class CountryController extends Controller
 
     public function get(Request $request, $id) : CountryResource
     {
+        $this->authorize("viewCountry", $this->_class);
+
         $language = $request->get("language", app()->getLocale());
         $country = $this->_service->find($id);
         return new CountryResource($country, $language);
@@ -43,6 +52,8 @@ class CountryController extends Controller
 
     public function update(Request $request, $id) : CountryResource
     {
+        $this->authorize("editCountry", $this->_class);
+
         $data = $request->all();
         $country = $this->_service->update($id, $data);
         return new CountryResource($country);
@@ -51,6 +62,8 @@ class CountryController extends Controller
 
     public function delete($id) : JsonResponse
     {
+        $this->authorize("deleteCountry", $this->_class);
+
         $this->_service->delete($id);
         return Response::ok();
     }
